Break up segments tests
