+++ test/core/parser/segments/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for the segments module."""
